requirements.txt
appdirs==1.4.4
asgiref==3.2.7
autoflake==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
django-nested-admin==3.3.2
django-registration-redux==2.8
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
filelock==3.0.12
flake8==3.8.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
mccabe==0.6.1
mysql-client==0.0.1
oauthlib==3.1.0
Pillow==7.2.0
pipenv==2020.8.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
PyMySQL==0.10.0
python-decouple==3.3
python-dotenv==0.17.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
zipp==3.1.0
